parameters are now upgraded with information to support postponed annotations
This is intended to occur when the signature is read from a function.
If it occurs in other places, we issue a warning because the data would
be incomplete.

We store a reference to the annotation contents as well as the function
it originally occurred on, which is needed to evaluate postponed
annotations (PEP-563).